live data converted to flow

+++ app/src/main/java/com/niyas/mishipay/repository/BarcodeRepository.kt
@@ -50,6 +50,6 @@ class BarcodeRepository {
     }
 
     fun getTotalAmountToPay() =
-        productsInCart.sumOf { it.price.toInt() * it.quantity }
+        productsInCart.sumOf { it.price * it.quantity }
 
 }
